requirements.txt
audioop-lts==0.2.1
discord.py==2.4.0
mcstatus==11.1.1
python-dotenv==1.0.1